user registration and login

+++ app/Http/Controllers/TasksController.php
@@ -15,8 +15,10 @@ class TasksController extends Controller
      */
     public function index()
     {
+        // メッセージ一覧を取得
         $tasks = Task::all();
-        
+
+        // メッセージ一覧ビューでそれを表示
         return view('tasks.index', [
             'tasks' => $tasks,
         ]);
@@ -45,16 +47,17 @@ class TasksController extends Controller
     public function store(Request $request)
     {
         $request->validate([
-            'status' => 'required | max:10'
+            'status' => 'required | max:10',
+            'content' => 'required | max:255'
             ]);
         
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
         
         return redirect('/');
-
     }
 
     /**
@@ -67,9 +70,15 @@ class TasksController extends Controller
     {
         $task = Task::findOrFail($id);
         
-        return view('tasks.show',[
-            'task' => $task,
-        ]);
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を編集
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show',[
+                'task' => $task,
+            ]);
+        }
+
+        return redirect('/');
+
     }
 
     /**
@@ -82,9 +91,14 @@ class TasksController extends Controller
     {
         $task = Task::findOrFail($id);
         
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を編集
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        }
+
+        return redirect('/');
     }
 
     /**
@@ -97,14 +111,17 @@ class TasksController extends Controller
     public function update(Request $request, $id)
     {
         $request->validate([
-            'status' => 'required | max:10'
-            ]);
+            'status' => 'required | max:10',
+            'content' => 'required | max:255'
+        ]);
         
         $task = Task::findOrFail($id);
+        
+        // メッセージを更新
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
-        
+
         return redirect('/');
     }
 
@@ -117,7 +134,11 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::findOrFail($id);
-        $task->delete();
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
         
         return redirect('/');
     }
